Includes Cristen's stuff
If you modify the path/name in main of cleanFile to point at a valid input file without comments, it will basically run a demo (read file, make token, display token, repeat). The displayToken function is just temporary. The displayError function includes a lot of potential internal errors and input errors we aren't required to check for but will probably make for easier debugging (and the extra ones can easily be removed).